// frontend/src/api/generated/model/criterionRead.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Call Analyzer Api
 * OpenAPI spec version: 0.1.0
 */
import type { CriterionReadDescription } from './criterionReadDescription';
import type { CriterionReadPrompt } from './criterionReadPrompt';
import type { CriterionAnswerType } from './criterionAnswerType';

export interface CriterionRead {
  template_id: number;
  name: string;
  description?: CriterionReadDescription;
  prompt?: CriterionReadPrompt;
  answer_type?: CriterionAnswerType;
  position?: number;
  id: number;
  created_at: string;
  updated_at: string;
}